requirements.txt
# Core API Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0

# LLM & Agent Orchestration
langchain==0.3.26
langchain-core==0.3.70
langchain-text-splitters==0.3.8
langgraph==0.5.4
langchain-google-genai==2.1.8

# Vector Database & Embeddings
weaviate-client==4.16.3
# Google Gemini API - EMBEDDINGS
google-genai>=1.0.0



# Document Processing
PyPDF2==3.0.1
python-multipart==0.0.20

# Configuration
python-dotenv==1.1.1
pydantic==2.11.7

# Data Processing
pandas>=2.0.0
numpy>=2.0.0

# Core utilities
typing-extensions
certifi
urllib3

# System monitoring
psutil

# Evaluation & Observability
ragas==0.1.19
langfuse==3.2.1

# Web Search
tavily-python==0.7.10

# Testing
pytest==8.2.0
pytest-asyncio==0.24.0
pytest-mock==3.12.0
httpx==0.26.0
pytest-cov==4.1.0
reportlab==4.0.9  # For generating test PDFs